Switch to Makie 0.24
The most recent release of the Makie family was a major breaking change,
introducing the new `ComputeGraph` architecture and eliminating
`MakieCore` as a separate package. This commit updates MolecularGraph to
be compatible with Makie 0.24, which includes:
- Creating a new extension package `MakieExt` to hold the code that
  previously leveraged `MakieCore`
- Updating the drawing recipes for the new API
